// src/services/rest-api/customer-service/model/combo-box-list-data-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.242
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { ComboBox } from './combo-box';

/**
 * 
 * @export
 * @interface ComboBoxListDataResponse
 */
export interface ComboBoxListDataResponse {
    /**
     * 
     * @type {number}
     * @memberof ComboBoxListDataResponse
     */
    'status'?: number;
    /**
     * 
     * @type {{ [key: string]: Array<string>; }}
     * @memberof ComboBoxListDataResponse
     */
    'errors'?: { [key: string]: Array<string>; } | null;
    /**
     * 
     * @type {string}
     * @memberof ComboBoxListDataResponse
     */
    'message'?: string | null;
    /**
     * 
     * @type {Array<ComboBox>}
     * @memberof ComboBoxListDataResponse
     */
    'data'?: Array<ComboBox> | null;
}